services/user-history/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic-settings==2.9.1
python-dotenv==1.1.0
tenacity==9.1.2
